+++ src/Resources/ResourceManager/Properties/Resources.Designer.cs
@@ -187,6 +187,15 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to CorrelationId: {0}.
+        /// </summary>
+        internal static string DeploymentCorrelationId {
+            get {
+                return ResourceManager.GetString("DeploymentCorrelationId", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Deployment &apos;{0}&apos; doesn&apos;t exist at management group &apos;{1}&apos;..
         /// </summary>
         internal static string DeploymentDoesntExistAtManagementGroupScope {
@@ -223,7 +232,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Error: Code={0}; Message={1}; Details={2}.
+        ///   Looks up a localized string similar to {0} (Code: {1});
+        ///{2};.
         /// </summary>
         internal static string DeploymentOperationErrorMessage {
             get {
@@ -232,7 +242,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Error: Code={0}; Message={1}.
+        ///   Looks up a localized string similar to {0} (Code:{1}).
         /// </summary>
         internal static string DeploymentOperationErrorMessageNoDetails {
             get {
@@ -241,16 +251,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to OperationId: {0}.
-        /// </summary>
-        internal static string DeploymentOperationId {
-            get {
-                return ResourceManager.GetString("DeploymentOperationId", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to Deployment &apos;{0}&apos; failed with error(s). Showing {1} out of {2} error(s)..
+        ///   Looks up a localized string similar to The deployment &apos;{0}&apos; failed. Showing {1} out of {2} error(s)..
         /// </summary>
         internal static string DeploymentOperationOuterError {
             get {
@@ -259,7 +260,16 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Target: ResourceType={0}; ResourceName={1}.
+        ///   Looks up a localized string similar to Status Message: {0}.
+        /// </summary>
+        internal static string DeploymentOperationStatusMessage {
+            get {
+                return ResourceManager.GetString("DeploymentOperationStatusMessage", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Target: {0}; {1}.
         /// </summary>
         internal static string DeploymentOperationTargetInfoInErrror {
             get {