+++ src/Resources/Resources/ChangeLog.md
@@ -18,8 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add message warning about view delay when creating a new Role Definition
 * Change policy cmdlets to output strongly-typed objects
-* Fiexed `Remove-AzResourceGroup -Id ResourceId`[#9882]
+* Fixed `Remove-AzResourceGroup -Id ResourceId`[#9882]
 
 ## Version 1.13.0
 * Fixed `Get-AzResource -ResourceGroupName -Name -ExpandProperties -ResourceType` to use actual apiVersion of resources instead of default apiVersion [#11267]
@@ -27,6 +28,14 @@
 * Small documentation change to `Get-AzResourceLock`. Added example.
 * Escaped single quote in parameter value of `Get-AzADUser` [#11317]
 * Added new cmdlets for Deployment Scripts (`Get-AzDeploymentScript`, `Get-AzDeploymentScriptLog`, `Save-AzDeploymentScriptLog`, `Remove-AzDeploymentScript`)
+* Updated nuget Microsoft.Azure.Management.ResourceManager to "3.4.0-preview"
+* Removed specific test cases as Error property in DeploymentValidateResult has changed to readonly from nuget 3.4.0-preview
+* Brought GenericResourceExpanded from SDK ResourceManager 3.4.0-preview
+* Added tag support for all Get cmdlets for deployment, as well as
+    - NewAzureManagementGroupDeploymentCmdlet
+    - NewAzureSubscriptionDeploymentCmdlet
+    - NewAzureResourceGroupDeploymentCmdlet
+    - NewAzureTenantDeploymentCmdlet
 
 ## Version 1.12.0
 * Fixed for null reference bug in `Get-AzRoleAssignment`